refactor: change files structure, update imports
