+++ src/app/index.ts
@@ -1,3 +1,3 @@
 export { ErrorBoundary } from './providers/ErrorBoundary'
-export { useAppDispatch, useAppSelector, wrapper } from './store/store'
-export type { RootState } from './store/store'
+export { wrapper } from './store/store'
+export type { AppDispatch, RootState } from './store/store'
