+++ src/entities/post/index.ts
@@ -9,5 +9,4 @@ export { PostPreviewCard } from './ui/PostPreviewCard/PostPreviewCard'
 export { PostsList } from './ui/PostsList/PostsList'
 export { PostsListSkeleton } from './ui/PostsListSeketon/PostsListSkeleton'
 export { PublicPostsList } from './ui/PublicPostsList/PublicPostsList'
-export { UserBanner } from './ui/UserBanner/UserBanner'
 export { ZoomIn } from './ui/ZoomIn/ZoomIn'
