+++ src/entities/user/index.ts
@@ -1,1 +1,2 @@
 export { CounterRegisteredUsers } from './ui/CounterRegisteredUsers/CounterRegisteredUsers'
+export { UserBanner } from './ui/UserBanner/UserBanner'
