+++ src/entities/user/ui/UserBanner/UserBanner.module.scss
