+++ src/entities/user/ui/UserBanner/UserBanner.tsx
