+++ src/pages/index.tsx
@@ -13,13 +13,13 @@ import { GetStaticPropsResult } from 'next'
 
 import s from './index.module.scss'
 
-type ErrorArgs = {
+type Props = {
   isPostsError?: boolean
   isUsersError?: boolean
 }
 
 export const getStaticProps = wrapper.getStaticProps(
-  store => async (): Promise<GetStaticPropsResult<ErrorArgs>> => {
+  store => async (): Promise<GetStaticPropsResult<Props>> => {
     const users = await store.dispatch(
       getTotalUsersCount.initiate(undefined, { forceRefetch: true })
     )
@@ -50,11 +50,6 @@ export const getStaticProps = wrapper.getStaticProps(
     }
   }
 )
-
-type Props = {
-  isPostsError?: boolean
-  isUsersError?: boolean
-}
 
 const Public: Page = ({ isPostsError, isUsersError }: Props) => {
   const { t } = useTranslation()
