+++ src/shared/hooks/index.ts
@@ -1,3 +1,5 @@
+export { useAppDispatch } from './useAppDispatch'
+export { useAppSelector } from './useAppSelector'
 export { useFormRevalidateWithLocale } from './useFormRevalidateWithLocale'
 export { useInfinityScroll } from './useInfinityScroll'
 export { useLoader } from './useLoader'
