+++ src/shared/hooks/useAppDispatch.ts
@@ -1,0 +1,5 @@
+import { useDispatch } from 'react-redux'
+
+import { AppDispatch } from '@/app'
+
+export const useAppDispatch: () => AppDispatch = useDispatch
