+++ src/shared/ui/Carousel/Carousel.tsx
@@ -3,11 +3,6 @@ import Image from 'next/image'
 import { Navigation, Pagination } from 'swiper/modules'
 import { Swiper, SwiperProps, SwiperSlide } from 'swiper/react'
 
-import './Carousel.scss'
-import 'swiper/css'
-import 'swiper/css/navigation'
-import 'swiper/css/pagination'
-
 export type ImagesUrlData = {
   url: string
 }
